remove bigdecimal jooq

+++ src/main/java/online/kheops/auth_server/album/AlbumResponseBuilder.java
@@ -4,7 +4,6 @@ import online.kheops.auth_server.entity.AlbumUser;
 import online.kheops.auth_server.user.UserResponse;
 import org.jooq.Record;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -48,7 +47,7 @@ public class AlbumResponseBuilder {
         this.numberOfStudies = (Integer) r.getValue("number_of_studies");
         this.numberOfSeries = (Integer) r.getValue("number_of_series");
         try {
-            this.numberOfInstances = ((BigDecimal) r.getValue("number_of_instances")).intValue();
+            this.numberOfInstances = (Integer) r.getValue("number_of_instances");
         } catch(NullPointerException e) {
             this.numberOfInstances = 0;
         }
@@ -79,7 +78,7 @@ public class AlbumResponseBuilder {
         this.numberOfStudies = (Integer) r.getValue("number_of_studies");
         this.numberOfSeries = (Integer) r.getValue("number_of_series");
         try {
-            this.numberOfInstances = ((BigDecimal) r.getValue("number_of_instances")).intValue();
+            this.numberOfInstances = (Integer) r.getValue("number_of_instances");
         } catch(NullPointerException e) {
             this.numberOfInstances = 0;
         }
@@ -89,7 +88,7 @@ public class AlbumResponseBuilder {
             this.modalities = new String[0];
         }
         try {
-            this.numberOfInstances = ((BigDecimal) r.getValue("number_of_instances")).intValue();
+            this.numberOfInstances = (Integer) r.getValue("number_of_instances");
         } catch(NullPointerException e) {
             this.numberOfInstances = 0;
         }
